pyproject.toml
[tool.poetry]
name = "fastapi_example"
version = "0.1.0"
description = "My Awesome Project!"
readme = "README.md"
authors = ["Author <author@example.com>"]
license = "MIT"
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = "^3.10"
dynaconf = "^3.1.9"
click = "^8.1.3"
SQLAlchemy = { extras = ["asyncio"], version = "^1.4.42" }
pydantic = "^1.10.2"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
aiosqlite = "^0.17.0"

[tool.poetry.plugins."scripts"]
fastapi_example = "fastapi_example.cmdline:main"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.20.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
__tablename__ = 'user'


[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(levelname)s %(name)s %(process)d %(thread)d %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"